Escape parentheses in MarkDown URLs
The Rebol-based "markymark" MarkDown parser was having some trouble
with parentheses in the already-parenthesized MarkDown links.  It's
probably for the best to just go ahead and enforce those as being
escaped, so the generator now does this.

It's a bit of an unfortunate choice that legal URL characters were
used in MarkDown to surround URL values...seems asking for problems.